+++ phpunit.xml.dist
@@ -12,6 +12,7 @@
     <testsuites>
         <testsuite name="Test Suite">
             <directory>tests</directory>
+            <directory>tests/SortedSets</directory>
         </testsuite>
     </testsuites>
     <filter>
